+++ libinitializer/CMakeLists.txt
@@ -22,6 +22,9 @@ endif()
 add_library(txpool_init TxPoolInitializer.cpp)
 target_link_libraries(txpool_init PUBLIC protocol_init tool txpool)
 
+add_library(mempool_init MemPoolInitializer.cpp)
+target_link_libraries(mempool_init PUBLIC mempool)
+
 add_library(ledger_init LedgerInitializer.cpp)
 target_link_libraries(ledger_init PRIVATE ledger)
 
@@ -29,7 +32,7 @@ add_library(baseline_init BaselineSchedulerInitializer.cpp GlobalStateStorageIni
 target_link_libraries(baseline_init transaction-scheduler transaction-executor storage)
 
 add_library(init Initializer.cpp)
-list(APPEND INIT_LIB_DEPENDS protocol_init baseline_init front_init txpool_init scheduler ledger_init storage executor rpc bcos-boostssl tomlplusplus::tomlplusplus)
+list(APPEND INIT_LIB_DEPENDS protocol_init baseline_init front_init txpool_init mempool_init scheduler ledger_init storage executor rpc bcos-boostssl tomlplusplus::tomlplusplus)
 if(WITH_LEDGER_ELECTION)
     list(APPEND INIT_LIB_DEPENDS leader_election)
 endif()
