+++ libinitializer/MemPoolInitializer.cpp
@@ -1,0 +1,6 @@
+#include "MemPoolInitializer.h"
+
+bcos::initializer::MemPoolInitializer::Ptr bcos::initializer::MemPoolInitializer::build()
+{
+    return std::make_shared<MemPoolInitializer>();
+}
