--- mempool/test/unittests/mempool/Web3TransactionsTest.cpp
@@ -1,540 +1,0 @@
-/**
- *  Copyright (C) 2025 FISCO BCOS.
- *  SPDX-License-Identifier: Apache-2.0
- */
-
-#include "bcos-mempool/Web3Transactions.h"
-#include "bcos-tars-protocol/protocol/TransactionImpl.h"
-#include <bcos-crypto/hash/Keccak256.h>
-#include <bcos-framework/ledger/EVMAccount.h>
-#include <bcos-framework/mempool/MemPool.h>
-#include <bcos-framework/transaction-executor/StateKey.h>
-#include <bcos-task/Wait.h>
-#include <bcos-utilities/Common.h>
-#include <boost/test/unit_test.hpp>
-#include <iterator>
-#include <range/v3/all.hpp>
-#include <unordered_map>
-
-using namespace bcos;
-using namespace bcos::txpool;
-using namespace bcos::protocol;
-using namespace bcos::crypto;
-using namespace bcos::executor_v1;
-
-namespace bcos::test
-{
-struct MapStateStorage
-{
-    using Value = storage::Entry;
-    std::unordered_map<std::string, std::unordered_map<std::string, Value>> data;
-    task::Task<std::optional<Value>> readOne(StateKeyView key)
-    {
-        auto [table, field] = key.get();
-        if (auto tIt = data.find(std::string(table)); tIt != data.end())
-        {
-            if (auto fIt = tIt->second.find(std::string(field)); fIt != tIt->second.end())
-            {
-                co_return std::make_optional(fIt->second);
-            }
-        }
-        co_return std::nullopt;
-    }
-
-    task::Task<std::optional<Value>> readOne(StateKey key)
-    {
-        co_return co_await readOne(StateKeyView{key});
-    }
-
-    template <class Keys>
-    task::Task<std::vector<std::optional<Value>>> readSome(Keys keys)
-    {
-        std::vector<std::optional<Value>> results;
-        if constexpr (::ranges::sized_range<Keys>)
-        {
-            results.reserve(::ranges::size(keys));
-        }
-        else
-        {
-            results.reserve(::ranges::distance(keys));
-        }
-
-        for (auto&& k : keys)
-        {
-            results.emplace_back(co_await readOne(StateKeyView{k}));
-        }
-        co_return results;
-    }
-
-    task::Task<void> writeOne(StateKey key, Value value)
-    {
-        StateKeyView view{key};
-        auto [table, field] = view.get();
-        data[std::string(table)][std::string(field)] = std::move(value);
-        co_return;
-    }
-
-    task::Task<void> writeOne(StateKeyView key, Value value)
-    {
-        auto [table, field] = key.get();
-        data[std::string(table)][std::string(field)] = std::move(value);
-        co_return;
-    }
-
-    template <class KVs>
-    task::Task<void> writeSome(KVs keyValues)
-    {
-        for (auto&& kv : keyValues)
-        {
-            StateKey key{std::get<0>(kv)};
-            auto& value = std::get<1>(kv);
-            StateKeyView view{key};
-            auto [table, field] = view.get();
-            data[std::string(table)][std::string(field)] = value;
-        }
-        co_return;
-    }
-
-    task::Task<bool> existsOne(StateKeyView key)
-    {
-        auto value = co_await readOne(key);
-        co_return value.has_value();
-    }
-
-    task::Task<bool> existsOne(StateKey key) { co_return co_await existsOne(StateKeyView{key}); }
-};
-
-static_assert(mempool::MemPool<Web3Transactions, MapStateStorage>);
-
-static bytes toBytes(std::string_view s)
-{
-    return {reinterpret_cast<const byte*>(s.data()),
-        reinterpret_cast<const byte*>(s.data()) + s.size()};
-}
-
-class TestTransactionImpl : public bcostars::protocol::TransactionImpl
-{
-public:
-    void markClean() { setTainted(false); }
-};
-
-static protocol::Transaction::Ptr makeTx(std::string_view senderBytes, int64_t nonce)
-{
-    auto tx = std::make_shared<TestTransactionImpl>();
-    tx->mutableInner().data.to.assign(senderBytes.begin(), senderBytes.end());
-    tx->setNonce(std::to_string(nonce));
-    tx->forceSender(toBytes(senderBytes));
-    Keccak256 hasher;
-    tx->calculateHash(hasher);
-    tx->markClean();
-    tx->setImportTime(nonce);
-    return tx;
-}
-
-static std::optional<std::string> readNonce(MapStateStorage& s, std::string_view sender)
-{
-    ledger::account::EVMAccount acc{s, sender, false};
-    return task::syncWait(acc.nonce());
-}
-
-static void setNonce(MapStateStorage& s, std::string_view sender, std::string nonce)
-{
-    ledger::account::EVMAccount acc{s, sender, false};
-    task::syncWait(acc.setNonce(std::move(nonce)));
-}
-
-BOOST_AUTO_TEST_SUITE(Web3TransactionsTest)
-
-BOOST_AUTO_TEST_CASE(seal_single_sender_contiguous)
-{
-    Web3Transactions pool;
-    std::string sender("aaaaaaaaaaaaaaaaaaaa", 20);
-    std::vector<protocol::Transaction::Ptr> txs;
-    txs.emplace_back(makeTx(sender, 0));
-    txs.emplace_back(makeTx(sender, 1));
-    txs.emplace_back(makeTx(sender, 2));
-    pool.add(txs);
-
-    MapStateStorage state{};
-    std::vector<protocol::Transaction::Ptr> out;
-    pool.seal(100, state, std::back_inserter(out));
-
-    BOOST_CHECK_EQUAL(out.size(), 3);
-    auto nonce = readNonce(state, sender);
-    BOOST_CHECK(nonce.has_value());
-    BOOST_CHECK_EQUAL(nonce.value(), "3");
-}
-
-BOOST_AUTO_TEST_CASE(seal_limit_within_single_sender)
-{
-    Web3Transactions pool;
-    constexpr int kSenderBytes = 20;
-    constexpr int kSealLimit = 2;
-    std::string sender("LLLLLLLLLLLLLLLLLLLL", kSenderBytes);
-
-    std::vector<protocol::Transaction::Ptr> txsA{
-        makeTx(sender, 0), makeTx(sender, 1), makeTx(sender, 2)};
-    pool.add(txsA);
-    std::string otherSender("MMMMMMMMMMMMMMMMMMMM", kSenderBytes);
-    pool.add(std::vector{makeTx(otherSender, 0)});
-
-    MapStateStorage state{};
-    std::vector<protocol::Transaction::Ptr> out;
-    pool.seal(kSealLimit, state, std::back_inserter(out));
-
-    BOOST_CHECK_EQUAL(out.size(), 2);
-    auto nonces = ::ranges::views::transform(out, [](auto& txPtr) {
-        return std::stoll(std::string(txPtr->nonce()));
-    }) | ::ranges::to<std::vector>();
-    ::ranges::sort(nonces);
-    BOOST_CHECK(nonces[0] == 0 && nonces[1] == 1);
-
-    auto nonceA = readNonce(state, sender);
-    BOOST_CHECK(nonceA.has_value());
-    BOOST_CHECK_EQUAL(nonceA.value(), "2");
-    auto nonceB = readNonce(state, otherSender);
-    BOOST_CHECK(!nonceB.has_value());
-}
-
-BOOST_AUTO_TEST_CASE(seal_limit_across_senders)
-{
-    Web3Transactions pool;
-    constexpr int kSenderBytes = 20;
-    constexpr int kSealLimit = 2;
-    std::string senderA("NNNNNNNNNNNNNNNNNNNN", kSenderBytes);
-    std::string senderB("OOOOOOOOOOOOOOOOOOOO", kSenderBytes);
-
-    pool.add(std::vector{makeTx(senderA, 0), makeTx(senderA, 2), makeTx(senderB, 0)});
-
-    MapStateStorage state{};
-    std::vector<protocol::Transaction::Ptr> out;
-    pool.seal(kSealLimit, state, std::back_inserter(out));
-
-    BOOST_CHECK_EQUAL(out.size(), 2);
-    bool foundA0 = false;
-    bool foundB0 = false;
-    for (auto& txPtr : out)
-    {
-        auto senderStr = std::string(txPtr->sender());
-        auto nonceVal = std::stoll(std::string(txPtr->nonce()));
-        if (senderStr == senderA && nonceVal == 0)
-        {
-            foundA0 = true;
-        }
-        if (senderStr == senderB && nonceVal == 0)
-        {
-            foundB0 = true;
-        }
-    }
-    BOOST_CHECK(foundA0 && foundB0);
-
-    auto nonceA = readNonce(state, senderA);
-    auto nonceB = readNonce(state, senderB);
-    BOOST_CHECK(nonceA.has_value());
-    BOOST_CHECK(nonceB.has_value());
-    BOOST_CHECK_EQUAL(nonceA.value(), "1");
-    BOOST_CHECK_EQUAL(nonceB.value(), "1");
-}
-
-BOOST_AUTO_TEST_CASE(seal_limit_exact_boundary)
-{
-    Web3Transactions pool;
-    constexpr int kSenderBytes = 20;
-    constexpr int kSealLimit = 2;
-    std::string sender("PPPPPPPPPPPPPPPPPPPP", kSenderBytes);
-
-    pool.add(std::vector{makeTx(sender, 0), makeTx(sender, 1), makeTx(sender, 2)});
-    MapStateStorage state{};
-    std::vector<protocol::Transaction::Ptr> out;
-    pool.seal(kSealLimit, state, std::back_inserter(out));
-
-    BOOST_CHECK_EQUAL(out.size(), 2);
-    auto nonces = ::ranges::views::transform(out, [](auto& txPtr) {
-        return std::stoll(std::string(txPtr->nonce()));
-    }) | ::ranges::to<std::vector>();
-    ::ranges::sort(nonces);
-    BOOST_CHECK(nonces[0] == 0 && nonces[1] == 1);
-
-    auto nonceA = readNonce(state, sender);
-    BOOST_CHECK(nonceA.has_value());
-    BOOST_CHECK_EQUAL(nonceA.value(), "2");
-}
-
-BOOST_AUTO_TEST_CASE(seal_multiple_senders_and_gaps)
-{
-    Web3Transactions pool;
-    constexpr int kSenderBytes = 20;
-    constexpr int kSealLimit = 100;
-    std::string senderAlpha("AAAAAAAAAAAAAAAAAAAA", kSenderBytes);
-    std::string senderBeta("BBBBBBBBBBBBBBBBBBBB", kSenderBytes);
-
-    pool.add(std::vector{makeTx(senderAlpha, 0), makeTx(senderAlpha, 2), makeTx(senderBeta, 0)});
-
-    MapStateStorage state{};
-    std::vector<protocol::Transaction::Ptr> out;
-    pool.seal(kSealLimit, state, std::back_inserter(out));
-
-    BOOST_CHECK_EQUAL(out.size(), 2);
-    auto sealedNoncesBySender = std::unordered_map<std::string, std::vector<int64_t>>{};
-    for (auto& txPtr : out)
-    {
-        sealedNoncesBySender[std::string(txPtr->sender())].push_back(
-            std::stoll(std::string(txPtr->nonce())));
-    }
-    BOOST_CHECK(sealedNoncesBySender.contains(senderAlpha));
-    BOOST_CHECK(sealedNoncesBySender.contains(senderBeta));
-    BOOST_CHECK(
-        sealedNoncesBySender[senderAlpha].size() == 1 && sealedNoncesBySender[senderAlpha][0] == 0);
-    BOOST_CHECK(
-        sealedNoncesBySender[senderBeta].size() == 1 && sealedNoncesBySender[senderBeta][0] == 0);
-
-    auto nonceA = readNonce(state, senderAlpha);
-    auto nonceB = readNonce(state, senderBeta);
-    BOOST_CHECK(nonceA.has_value());
-    BOOST_CHECK(nonceB.has_value());
-    BOOST_CHECK_EQUAL(nonceA.value(), "1");
-    BOOST_CHECK_EQUAL(nonceB.value(), "1");
-}
-
-BOOST_AUTO_TEST_CASE(seal_with_existing_ledger_nonce)
-{
-    Web3Transactions pool;
-    constexpr int kSenderBytes2 = 20;
-    constexpr int kSealLimit2 = 100;
-    std::string sender("CCCCCCCCCCCCCCCCCCCC", kSenderBytes2);
-
-    MapStateStorage state{};
-    setNonce(state, sender, "5");
-
-    constexpr int64_t kNonce3 = 3;
-    constexpr int64_t kNonce4 = 4;
-    constexpr int64_t kNonce5 = 5;
-    constexpr int64_t kNonce6 = 6;
-    pool.add(std::vector{makeTx(sender, kNonce3), makeTx(sender, kNonce4), makeTx(sender, kNonce5),
-        makeTx(sender, kNonce6)});
-
-    std::vector<protocol::Transaction::Ptr> out;
-    pool.seal(kSealLimit2, state, std::back_inserter(out));
-
-    BOOST_CHECK_EQUAL(out.size(), 2);
-    auto nonceVec = std::vector<int64_t>{
-        std::stoll(std::string(out[0]->nonce())), std::stoll(std::string(out[1]->nonce()))};
-    ::ranges::sort(nonceVec);
-    BOOST_CHECK(nonceVec[0] == 5 && nonceVec[1] == 6);
-
-    auto nonce = readNonce(state, sender);
-    BOOST_CHECK(nonce.has_value());
-    BOOST_CHECK_EQUAL(nonce.value(), "7");
-}
-
-BOOST_AUTO_TEST_CASE(remove_by_state_drops_confirmed)
-{
-    Web3Transactions pool;
-    constexpr int kSenderBytes = 20;
-    constexpr int kSealLimit = 100;
-    std::string sender("DDDDDDDDDDDDDDDDDDDD", kSenderBytes);
-
-    std::vector<protocol::Transaction::Ptr> txs;
-    constexpr int kMaxNonce = 5;
-    for (int i = 0; i <= kMaxNonce; ++i)
-    {
-        txs.emplace_back(makeTx(sender, i));
-    }
-    pool.add(txs);
-
-    MapStateStorage state{};
-    setNonce(state, sender, "3");
-    pool.remove(state);
-
-    setNonce(state, sender, "4");
-    std::vector<protocol::Transaction::Ptr> out;
-    pool.seal(kSealLimit, state, std::back_inserter(out));
-
-    BOOST_CHECK_EQUAL(out.size(), 2);
-    auto firstNonce = std::stoll(std::string(out[0]->nonce()));
-    auto secondNonce = std::stoll(std::string(out[1]->nonce()));
-    auto nonceArray = std::array<long, 2>{firstNonce, secondNonce};
-    ::ranges::sort(nonceArray);
-    BOOST_CHECK(nonceArray[0] == 4 && nonceArray[1] == 5);
-}
-
-BOOST_AUTO_TEST_CASE(remove_by_hashes_respects_per_sender_max)
-{
-    Web3Transactions pool;
-    constexpr int kSenderBytes = 20;
-    constexpr int kSealLimit = 100;
-    std::string senderAName("EEEEEEEEEEEEEEEEEEEE", kSenderBytes);
-    std::string senderBName("FFFFFFFFFFFFFFFFFFFF", kSenderBytes);
-
-    std::vector<protocol::Transaction::Ptr> txsA;
-    std::vector<protocol::Transaction::Ptr> txsB;
-    constexpr int kMaxANonce = 5;
-    constexpr int kMaxBNonce = 2;
-    for (int i = 0; i <= kMaxANonce; ++i)
-    {
-        txsA.emplace_back(makeTx(senderAName, i));
-    }
-    for (int i = 0; i <= kMaxBNonce; ++i)
-    {
-        txsB.emplace_back(makeTx(senderBName, i));
-    }
-    pool.add(txsA);
-    pool.add(txsB);
-
-    std::vector<crypto::HashType> toRemove{
-        txsA[1]->hash(), txsA[3]->hash(), txsB[0]->hash(), txsB[1]->hash()};
-    crypto::HashListView view{toRemove};
-    pool.remove(view);
-
-    MapStateStorage state{};
-    setNonce(state, senderAName, "4");
-    setNonce(state, senderBName, "2");
-
-    std::vector<protocol::Transaction::Ptr> out;
-    pool.seal(kSealLimit, state, std::back_inserter(out));
-
-    BOOST_CHECK_EQUAL(out.size(), 3);
-    auto got = ::ranges::views::transform(out, [](auto& txPtr) {
-        return std::make_pair(
-            std::string(txPtr->sender()), std::stoll(std::string(txPtr->nonce())));
-    }) | ::ranges::to<std::vector>();
-    int aCount = 0;
-    int bCount = 0;
-    bool aHas4 = false;
-    bool aHas5 = false;
-    bool bHas2 = false;
-    constexpr int kNonce4 = 4;
-    constexpr int kNonce5 = 5;
-    constexpr int kNonce2 = 2;
-    for (auto& [senderStr, nonceValue] : got)
-    {
-        if (senderStr == senderAName)
-        {
-            ++aCount;
-            if (nonceValue == kNonce4)
-            {
-                aHas4 = true;
-            }
-            if (nonceValue == kNonce5)
-            {
-                aHas5 = true;
-            }
-        }
-        if (senderStr == senderBName)
-        {
-            ++bCount;
-            if (nonceValue == kNonce2)
-            {
-                bHas2 = true;
-            }
-        }
-    }
-    BOOST_CHECK_EQUAL(aCount, 2);
-    BOOST_CHECK_EQUAL(bCount, 1);
-    BOOST_CHECK(aHas4 && aHas5 && bHas2);
-}
-
-BOOST_AUTO_TEST_CASE(get_returns_in_order_with_null_for_missing)
-{
-    Web3Transactions pool;
-    constexpr int kSenderBytes = 20;
-    std::string senderA("GGGGGGGGGGGGGGGGGGGG", kSenderBytes);
-    std::string senderB("HHHHHHHHHHHHHHHHHHHH", kSenderBytes);
-
-    auto txA0 = makeTx(senderA, 0);
-    auto txB1 = makeTx(senderB, 1);
-    pool.add(std::vector{txA0, txB1});
-
-    constexpr int64_t kMissingNonce = 9;
-    std::string missingSender("IIIIIIIIIIIIIIIIIIII", kSenderBytes);
-    auto missingTx = makeTx(missingSender, kMissingNonce);
-    auto missingHash = missingTx->hash();
-
-    std::vector<crypto::HashType> query{txB1->hash(), missingHash, txA0->hash()};
-    auto results = pool.get(query);
-
-    BOOST_CHECK_EQUAL(results.size(), query.size());
-    BOOST_CHECK(results[0]);
-    BOOST_CHECK_EQUAL(results[0]->hash(), txB1->hash());
-    BOOST_CHECK(!results[1]);
-    BOOST_CHECK(results[2]);
-    BOOST_CHECK_EQUAL(results[2]->hash(), txA0->hash());
-}
-
-BOOST_AUTO_TEST_CASE(get_duplicate_hashes_returns_duplicates)
-{
-    Web3Transactions pool;
-    constexpr int kSenderBytes = 20;
-    constexpr int64_t kNonce = 7;
-    std::string sender("JJJJJJJJJJJJJJJJJJJJ", kSenderBytes);
-    auto tx = makeTx(sender, kNonce);
-    pool.add(std::vector{tx});
-
-    std::vector<crypto::HashType> query{tx->hash(), tx->hash()};
-    auto results = pool.get(query);
-
-    BOOST_CHECK_EQUAL(results.size(), 2);
-    BOOST_CHECK(results[0]);
-    BOOST_CHECK(results[1]);
-    BOOST_CHECK_EQUAL(results[0]->hash(), tx->hash());
-    BOOST_CHECK_EQUAL(results[1]->hash(), tx->hash());
-}
-
-BOOST_AUTO_TEST_CASE(get_empty_input_returns_empty)
-{
-    Web3Transactions pool;
-    std::vector<crypto::HashType> emptyQuery;
-    auto results = pool.get(emptyQuery);
-    BOOST_CHECK(results.empty());
-}
-
-BOOST_AUTO_TEST_CASE(testAddNullTransaction)
-{
-    Web3Transactions pool;
-    std::vector<protocol::Transaction::Ptr> txs{nullptr};
-    BOOST_CHECK_NO_THROW(pool.add(txs));
-}
-
-BOOST_AUTO_TEST_CASE(testAddEmptyHashTransaction)
-{
-    Web3Transactions pool;
-    auto tx = std::make_shared<TestTransactionImpl>();
-    tx->setNonce("0");
-    tx->forceSender(toBytes("aaaaaaaaaaaaaaaaaaaa"));
-    tx->markClean();
-    std::vector<protocol::Transaction::Ptr> txs{tx};
-    BOOST_CHECK_NO_THROW(pool.add(txs));
-}
-
-BOOST_AUTO_TEST_CASE(testAddInvalidNonceTransaction)
-{
-    Web3Transactions pool;
-    auto tx = std::make_shared<TestTransactionImpl>();
-    tx->setNonce("not_a_number");
-    tx->forceSender(toBytes("bbbbbbbbbbbbbbbbbbbb"));
-    Keccak256 hasher;
-    tx->calculateHash(hasher);
-    tx->markClean();
-    std::vector<protocol::Transaction::Ptr> txs{tx};
-    BOOST_CHECK_NO_THROW(pool.add(txs));
-}
-
-BOOST_AUTO_TEST_CASE(testAddTaintedTransaction)
-{
-    Web3Transactions pool;
-    auto taintedTx = std::make_shared<TestTransactionImpl>();
-    taintedTx->setNonce("0");
-    taintedTx->forceSender(toBytes("cccccccccccccccccccc"));
-    Keccak256 hasher;
-    taintedTx->calculateHash(hasher);
-
-    auto taintedHash = taintedTx->hash();
-    BOOST_CHECK_THROW(pool.add(std::vector<protocol::Transaction::Ptr>{taintedTx}), bcos::Exception);
-
-    auto result = pool.get(std::vector<bcos::crypto::HashType>{taintedHash});
-    BOOST_CHECK_EQUAL(result.size(), 1);
-    BOOST_CHECK(!result[0]);
-}
-
-BOOST_AUTO_TEST_SUITE_END()
-}  // namespace bcos::test
